Roll back on sass-rails version
Why? Because we need a newer version of sass. Though it does not
make sense, the older sass-rails gem had a looser version specification
for sass, and so by rolling back sass-rails we allow bundler to choose
a newer version of sass itself.

sass 3.3.x is required because sass 3.2.x fails to parse MarkUs scss code.